services/osm-importer/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
shapely>=2.0.0
tqdm>=4.66.0
